Move URIUtils functionality into backend.javascript
It is only used there at this point.
